Move stringify and parse functions to external file and update import path in app.js
